+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Next version
+
+- Add checkboxes_dropdown helper
+- Add FilterStrategyByKeyValues
+- bs_button improvements
+- Add Filter#any? and Filter#empty?
+- Add i18n actions
+
 ## 1.0.5
 
 - Fix `import_button` helper